// Docker/Docker_Service_Face/ai-projects/oryza-ai-frontend/interfaces/camera.ts
export interface CameraRes {
  name: string;
  ip_address: string;
  port: number;
  username: string;
  password: string;
  rtsp: string;
  rtsp_vms: string;
  is_ai: boolean;
  other_info: any;
  id: string;
  company: string;
  brand_camera: any;
  created: Date;
  modified: Date;
  ward_id?: string;
}